Move mikeio1d.open() into top_level module
