+++ mikeio1d/__init__.py
@@ -47,6 +47,7 @@ clr.AddReference("DHI.Mike1D.MikeIO")
 
 from .res1d import Res1D
 from .xns11 import Xns11
+from .top_level import open
 
 from .various import allow_nested_autocompletion_for_ipython
 
@@ -54,48 +55,4 @@ allow_nested_autocompletion_for_ipython(Res1D)
 allow_nested_autocompletion_for_ipython(Xns11)
 
 
-def open(file_name: str | Path, **kwargs) -> Res1D | Xns11:
-    """Open a file type supported by MIKE IO 1D file.
-
-    Parameters
-    ----------
-    file_name : str or Path
-        Path to the file to read.
-    **kwargs
-        Additional keyword arguments to pass to the relevant constructor.
-
-    See Also
-    --------
-    mikeio1d.Res1D
-    mikeio1d.Xns11
-
-    Returns
-    -------
-    Res1D or Xns11
-        The object representing the 1D file.
-
-    Examples
-    --------
-    >>> import mikeio1d
-    >>> res = mikeio1d.open("results.res1d")
-    >>> res.nodes.read()
-
-    >>> xs = mikeio1d.open("cross_section.xns11")
-    >>> xs
-    """
-    if isinstance(file_name, str):
-        file_name = Path(file_name)
-
-    if not file_name.exists():
-        raise FileNotFoundError(f"File not found: {file_name}")
-
-    suffix = file_name.suffix.lower()
-    file_name = str(file_name)
-
-    if suffix in Res1D.get_supported_file_extensions():
-        return Res1D(file_name, **kwargs)
-    elif suffix in Xns11.get_supported_file_extensions():
-        return Xns11(file_name, **kwargs)
-
-
-__all__ = ["Res1D", "Xns11"]
+__all__ = ["Res1D", "Xns11", "open"]
